[UDATE] REparando modulo escanero y actualizando el de usuario
