+++ AuditSentinel/AuditSentinel/Pages/Registrar/Editar.cshtml
@@ -12,16 +12,17 @@
     <input type="hidden" asp-for="UserId" />
 
     <div class="mb-3">
+        <label asp-for="Input.UserName">Nombre</label>
+        <input asp-for="Input.UserName" class="form-control" />
+        <span asp-validation-for="Input.UserName" class="text-danger"></span>
+    </div>
+    <div class="mb-3">
         <label asp-for="Input.Email"></label>
         <input asp-for="Input.Email" class="form-control" />
         <span asp-validation-for="Input.Email" class="text-danger"></span>
     </div>
 
-    <div class="mb-3">
-        <label asp-for="Input.UserName"></label>
-        <input asp-for="Input.UserName" class="form-control" />
-        <span asp-validation-for="Input.UserName" class="text-danger"></span>
-    </div>
+    
 
     <div class="mb-3">
         <label>Roles</label>
